Version 0.2.01

+++ dbd_project/database/migrations/2018_08_15_225423_create_actividad_table.php
@@ -22,7 +22,7 @@ class CreateActividadTable extends Migration
             $table->string('pais');
             $table->string('ciudad');
             $table->string('calle');
-            $table->smallInteger('valor_entrada');
+            $table->integer('valor_entrada');
             $table->integer('cupos');
             $table->integer('descuento');
             $table->timestamps();
